requirements.txt
opencv-python
numpy
pix2tex
torch
Pillow
Flask
Werkzeug
matplotlib

